+++ main.py
@@ -1,5 +1,5 @@
 from system import account
 
-print("RUNE OS ALPHA 3.3")
+print("RUNE OS ALPHA 3.4")
 print("Booting up...")
 account.login()