+++ system/account/__init__.py
@@ -12,10 +12,10 @@ DEFAULT_JSON_DATA = {"CURRENT": None, "ACCOUNTS": []}
 def update():
     global json_data
     try:
-        with open("info.json") as f:
+        with open("system/info.json") as f:
             json_data = json.loads(f.read())
     except FileNotFoundError:
-        with open("info.json", "w+") as f:
+        with open("system/info.json", "w+") as f:
             f.write(json.dumps(DEFAULT_JSON_DATA))
             json_data = DEFAULT_JSON_DATA
 
@@ -26,7 +26,7 @@ update()
 
 
 def save():
-    with open("info.json", "w") as f:
+    with open("system/info.json", "w") as f:
         f.write(json.dumps(json_data))